+++ crates/radicle-cli/src/lib.rs
@@ -8,6 +8,7 @@ pub mod pager;
 pub mod project;
 pub mod terminal;
 
+mod common_args;
 mod warning;
 
 extern crate radicle_localtime as localtime;